requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
cloudinary==1.29.0
dj-database-url==0.5.0
dj-places==5.1.0
Django==4.0.5
django-bootstrap4==22.1
django-braces==1.15.0
django-crispy-forms==1.14.0
django-google-maps==0.13.0
django-heroku==0.3.1
django-material==1.10.0
django-model-utils==4.2.0
django-star-ratings==0.9.2
fontawesomefree==6.1.1
gunicorn==20.1.0
psycopg2==2.9.3
python-decouple==3.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.9
whitenoise==6.2.0
